requirements.txt
playwright>=1.33.0
pyyaml>=6.0 